Gestion fournisseurs2

+++ ProjetPortailFournisseurs/routes/web.php
@@ -6,6 +6,10 @@ use App\Http\Controllers\PostController;
 use App\Http\Controllers\UsagersController;
 use App\Http\Controllers\FournisseursController;
 use App\Http\Controllers\AdminsController;
+use App\Http\Controllers\ResponsablesController;
+use App\Livewire\Recherche;
+use App\Http\Middleware\IsAdmin;
+use App\Http\Middleware\IsAuthorized;
 
 // Route pour Usagers
 
@@ -15,6 +19,7 @@ Route::get('/usagers/connexion',
 Route::post('/LoginUsagers', 
 [UsagersController::class, 'login'])->name('Usagers.login');
 
+Route::post('/logout', [UsagersController::class, 'logout'])->name('logout');
 
 
 Route::get('/', function () {
@@ -66,29 +71,41 @@ Route::patch('/fournisseur/{fournisseur}/modification',
 [FournisseursController::class, 'update'])->name('Fournisseurs.update');
 
 
+// Routes pour responsables et commis
+
+Route::middleware([IsAuthorized::class])->group(function () {
+    Route::get('/Recherche', Recherche::class)->name('fournisseurs.recherche');
+
+    Route::get('/fournisseurs/selection', [FournisseursController::class, 'showSelected'])->name('fournisseurs.selectionnes');
+
+    Route::get('/fournisseurs/{id}', [FournisseursController::class, 'showFiche'])->name('fournisseurs.showFiche');
+
+    Route::get('/fournisseurs/{id}/historique', [FournisseursController::class, 'showHistorique'])->name('fournisseurs.historique');
+
+    Route::get('/fournisseurs/{id}/edit', [FournisseursController::class, 'editFiche'])->name('fournisseurs.editFiche');
+
+    Route::put('/fournisseurs/{id}/modifier', [FournisseursController::class, 'modifierFournisseur'])->name('fournisseurs.modifierFournisseur');
+    
+    Route::put('/fournisseurs/{id}/modifieretat', [FournisseursController::class, 'modifierEtatFournisseur'])->name('fournisseurs.modifierEtatFournisseur');
+
+});
 
 // Routes admin
-Route::get('/admin', 
-[AdminsController::class, 'index'])->name('Admins.Panel');
 
-Route::get('/admin/usagers', 
-[AdminsController::class, 'gestionUsagers'])->name('Admins.Usagers');
-
-Route::get('/admin/usagers/nouveau', 
-[AdminsController::class, 'createUser'])->name('Admins.Usagers.Creation');
-
-Route::post('/admin/usagers/add', 
-[AdminsController::class,'storeUser'])->name('Admin.Usager.Store');
-
-Route::get('/admin/parametres', 
-[AdminsController::class, 'parametres'])->name('Admins.Parametres');
-
-Route::get('/admin/courriel', 
-[AdminsController::class, 'modelesCourriel'])->name('Admins.Courriel');
-
-Route::delete('/admin/usagers/{id}',
-[UsagersController::class,'destroy'])->name('Admins.Usager.Supprimer');
-
-Route::put('/admin/usagers/{id}/update-role',
-[AdminsController::class, 'updateRole'])->name('Admins.Usager.UpdateRole');
-
+Route::middleware([IsAdmin::class])->group(function () {
+    Route::get('/admin', [AdminsController::class, 'index'])->name('Admins.Panel');
+    
+    Route::get('/admin/usagers', [AdminsController::class, 'gestionUsagers'])->name('Admins.Usagers');
+    
+    Route::get('/admin/usagers/nouveau', [AdminsController::class, 'createUser'])->name('Admins.Usagers.Creation');
+    
+    Route::post('/admin/usagers/add', [AdminsController::class,'storeUser'])->name('Admin.Usager.Store');
+    
+    Route::get('/admin/parametres', [AdminsController::class, 'parametres'])->name('Admins.Parametres');
+    
+    Route::get('/admin/courriel', [AdminsController::class, 'modelesCourriel'])->name('Admins.Courriel');
+    
+    Route::delete('/admin/usagers/{id}', [UsagersController::class,'destroy'])->name('Admins.Usager.Supprimer');
+    
+    Route::put('/admin/usagers/{id}/update-role', [AdminsController::class, 'updateRole'])->name('Admins.Usager.UpdateRole');    
+});
